ci: extract shared logic between auto-merge and nudge stale workflows
