+++ src/api/autoMergeStale.ts
@@ -2,11 +2,9 @@ import {execFile as execFileCb} from 'node:child_process'
 import {promisify} from 'node:util'
 
 import {outdent} from 'outdent'
-import {ZodError} from 'zod'
-import {fromZodError} from 'zod-validation-error'
 
-import {githubPrListSchema} from '../schemas'
 import type {GitHubPRList, Locale} from '../types'
+import {getPendingAutoTranslatedPRs} from '../util/getPendingAutoTranslatedPRs'
 import {getRootPath} from '../util/getRootPath'
 import {AUTO_TRANSLATE_BRANCH_PREFIX} from './autoTranslate'
 import {PR_LABEL_AWAITING_REVIEW} from './ghLabels'
@@ -22,8 +20,6 @@ const execFile = promisify(execFileCb)
 export const STALE_MERGE_THRESHOLD_DAYS = 14
 
 const ONE_DAY_MS = 864e5
-const GH_REPO_OWNER = 'sanity-io'
-const GH_REPO_NAME = 'locales'
 
 /**
  * Options for the auto merge stale operation
@@ -65,7 +61,7 @@ export async function autoMergeStale(options: AutoMergeStaleOptions): Promise<vo
   logger('Finding pending auto-translated PRs')
   const [locales, pendingPRs] = await Promise.all([
     getLocaleRegistry(),
-    findPendingMergeableAutoTranslatedPRs(),
+    getPendingAutoTranslatedPRs(),
   ])
 
   // Filter down to the PRs that are _stale_, eg beyond our threshold
@@ -114,44 +110,6 @@ async function commentAndMergeStalePR(
   await execFile('gh', ['pr', 'comment', `${pr.number}`, '--body', comment], execOptions)
   await execFile('gh', ['pr', 'edit', `${pr.number}`, '--remove-label', removeLabel], execOptions)
   await execFile('gh', ['pr', 'merge', `${pr.number}`, '--squash', '--delete-branch'], execOptions)
-}
-
-async function findPendingMergeableAutoTranslatedPRs() {
-  const rootPath = await getRootPath()
-  const {stdout} = await execFile(
-    'gh',
-    [
-      'pr',
-      'list',
-      '--label',
-      PR_LABEL_AWAITING_REVIEW,
-      '--json',
-      'createdAt,mergeable,number,headRefName,headRepository,headRepositoryOwner,labels',
-    ],
-    {
-      cwd: rootPath,
-      // eslint-disable-next-line no-process-env
-      env: {...process.env, CLICOLOR: '0'},
-    },
-  )
-
-  let pending: GitHubPRList = []
-  try {
-    pending = githubPrListSchema.parse(JSON.parse(stdout))
-  } catch (err: unknown) {
-    throw err instanceof ZodError ? fromZodError(err) : err
-  }
-
-  return pending.filter(
-    (pr) =>
-      // PR is in our own repository
-      pr.headRepositoryOwner.login === GH_REPO_OWNER &&
-      pr.headRepository.name === GH_REPO_NAME &&
-      // PR is an auto-translated PR
-      pr.headRefName.startsWith(`${AUTO_TRANSLATE_BRANCH_PREFIX}/`) &&
-      // Does not have any conflicts
-      pr.mergeable === 'MERGEABLE',
-  )
 }
 
 function getCommentBody(locale: Locale) {
